--- src/sound_test/java_sound/AudioPlayerExample2.java
@@ -1,77 +1,0 @@
-package sound_test.java_sound;
-
-import java.io.File;
-import java.io.IOException;
-
-import javax.sound.sampled.AudioFormat;
-import javax.sound.sampled.AudioInputStream;
-import javax.sound.sampled.AudioSystem;
-import javax.sound.sampled.DataLine;
-import javax.sound.sampled.LineUnavailableException;
-import javax.sound.sampled.SourceDataLine;
-import javax.sound.sampled.UnsupportedAudioFileException;
-
-/**
- * This is an example program that demonstrates how to play back an audio file
- * using the SourceDataLine in Java Sound API.
- * @author www.codejava.net
- *
- */
-public class AudioPlayerExample2 {
-
-    // size of the byte buffer used to read/write the audio stream
-    private static final int BUFFER_SIZE = 4096;
-
-    /**
-     * Play a given audio file.
-     * @param audioFilePath Path of the audio file.
-     */
-    void play(String audioFilePath) {
-        File audioFile = new File(audioFilePath);
-        try {
-            AudioInputStream audioStream = AudioSystem.getAudioInputStream(audioFile);
-
-            AudioFormat format = audioStream.getFormat();
-
-            DataLine.Info info = new DataLine.Info(SourceDataLine.class, format);
-
-            SourceDataLine audioLine = (SourceDataLine) AudioSystem.getLine(info);
-
-            audioLine.open(format);
-
-            audioLine.start();
-
-            System.out.println("Playback started.");
-
-            byte[] bytesBuffer = new byte[BUFFER_SIZE];
-            int bytesRead = -1;
-
-            while ((bytesRead = audioStream.read(bytesBuffer)) != -1) {
-                audioLine.write(bytesBuffer, 0, bytesRead);
-            }
-
-            audioLine.drain();
-            audioLine.close();
-            audioStream.close();
-
-            System.out.println("Playback completed.");
-
-        } catch (UnsupportedAudioFileException ex) {
-            System.out.println("The specified audio file is not supported.");
-            ex.printStackTrace();
-        } catch (LineUnavailableException ex) {
-            System.out.println("Audio line for playing back is unavailable.");
-            ex.printStackTrace();
-        } catch (IOException ex) {
-            System.out.println("Error playing the audio file.");
-            ex.printStackTrace();
-        }
-    }
-
-    public static void main(String[] args) {
-        String audioFilePath = "src/Dosed.wav";
-        AudioPlayerExample2 player = new AudioPlayerExample2();
-        player.play(audioFilePath);
-    }
-
-}
